off-chain-code/AKEB/requirements.txt
asgiref==3.7.2
# backports.zoneinfo==0.2.1
# backports is compatible with Python version < 3.9
backports.zoneinfo;python_version<"3.9"
decouple==0.0.7
Django==4.2.2
djangorestframework==3.14.0
pytz==2023.3
sqlparse==0.4.4
typing_extensions==4.6.3
pycryptodome==3.18.0